requirements.txt
# PlantMind - Python Requirements
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Calendar Integration
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
caldav==1.3.9
icalendar==5.0.11

# AI & ML
google-generativeai==0.3.2
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
apscheduler==3.10.4
